Moving test.proto to an outer directory so others can depend on it.
