+++ test/cpp/interop/server.cc
@@ -48,7 +48,7 @@
 #include <grpc++/server_credentials.h>
 #include <grpc++/status.h>
 #include <grpc++/stream.h>
-#include "test/cpp/interop/test.grpc.pb.h"
+#include "test/proto/test.grpc.pb.h"
 #include "test/cpp/interop/empty.grpc.pb.h"
 #include "test/cpp/interop/messages.grpc.pb.h"
 #include "test/cpp/interop/server_helper.h"
